requirements.txt
sentence-transformers==2.7.0
faiss-cpu>=1.9.0
rank-bm25==0.2.2
httpx>=0.27.0
numpy>=2.0.0
python-dotenv==1.0.1
torch>=2.6.0
transformers>=4.41.1
scikit-learn>=1.4.2
